refactor(onchain-interactions): Revert moving `startPrometheusServer` to `base-utils`
`base-utils` is our shared base library used by frontend, CLI and backend projects.
`prometheus` and `express` are backend-only dependencies that should only be imported by `apps/`.
