+++ libs/ts/base-utils/package.json
@@ -71,11 +71,6 @@
       "require": "./dist/cjs/type-level.cjs",
       "import": "./dist/esm/type-level.mjs",
       "types": "./src/type-level.ts"
-    },
-    "./prometheus": {
-      "require": "./dist/cjs/prometheus.cjs",
-      "import": "./dist/esm/prometheus.mjs",
-      "types": "./src/prometheus.ts"
     }
   },
   "scripts": {
@@ -84,7 +79,6 @@
     "test": "yarn vitest --typecheck -w false"
   },
   "devDependencies": {
-    "@types/express": "^5",
     "@types/node": "^22.10.2",
     "glob": "^11.0.0",
     "tsx": "^4.19.2",
@@ -92,8 +86,6 @@
     "vitest": "^3.0.2"
   },
   "dependencies": {
-    "effect": "^3.12.1",
-    "express": "^5.1.0",
-    "prom-client": "^15.1.3"
+    "effect": "^3.12.1"
   }
 }
