+++ libs/ts/onchain-interactions/package.json
@@ -14,12 +14,14 @@
     "chalk": "^5.3.0",
     "chalk-template": "^1.1.0",
     "dotenv": "^16.4.7",
+    "express": "^5.1.0",
     "prom-client": "^15.1.3",
     "web3": "^4.15.0",
     "yargs": "^17.7.2"
   },
   "devDependencies": {
     "@blocksense/base-utils": "workspace:*",
+    "@types/express": "^5",
     "@types/yargs": "^17.0.33",
     "tsx": "^4.19.2"
   }
