+++ libs/ts/onchain-interactions/src/scripts/balance-script.ts
@@ -2,7 +2,6 @@ import Web3 from 'web3';
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
 import chalk from 'chalk';
-import client from 'prom-client';
 import {
   getOptionalRpcUrl,
   networkMetadata,
@@ -10,7 +9,8 @@ import {
 } from '@blocksense/base-utils/evm';
 import { deployedNetworks } from '../types';
 import { getEnvStringNotAssert } from '@blocksense/base-utils/env';
-import { startPrometheusServer } from '@blocksense/base-utils/prometheus';
+import client from 'prom-client';
+import express from 'express';
 
 const balanceGauge = new client.Gauge({
   name: 'eth_account_balance',
@@ -21,6 +21,20 @@ const balanceGauge = new client.Gauge({
 function filterSmallBalance(balance: string, threshold = 1e-6): number {
   return Number(balance) < threshold ? 0 : Number(balance);
 }
+const startPrometheusServer = (host: string, port: number): void => {
+  const app = express();
+  app.get('/metrics', async (_req, res) => {
+    res.set('Content-Type', client.register.contentType);
+    res.end(await client.register.metrics());
+  });
+  app.listen(port, host, () => {
+    console.log(
+      chalk.blue(
+        `Prometheus metrics exposed at http://${host}:${port}/metrics`,
+      ),
+    );
+  });
+};
 
 const main = async (): Promise<void> => {
   const sequencerAddress = getEnvStringNotAssert('SEQUENCER_ADDRESS');
